requirements.txt
streamlit
torch
torchvision
timm==0.9.2
pillow
numpy
matplotlib
grad-cam
opencv-python-headless
fpdf2